+++ src/services/contract/types.ts
@@ -4,6 +4,6 @@ export type MonetaryContractJson = typeof monetaryCelo | typeof monetaryAlfajore
 export type GovernanceContractJson = typeof governanceCelo | typeof governanceAlfajores;
 export type ContractJson = MonetaryContractJson | GovernanceContractJson;
 export type Shared = typeof governanceShared | typeof monetaryShared;
-export type Contracts = ContractJson['contracts'];
+export type ContractsJson = ContractJson['contracts'];
 export type MonetaryContracts = Extract<keyof MonetaryContractJson['contracts'], string>;
 export type GovernanceContracts = Extract<keyof GovernanceContractJson['contracts'], string>;
