+++ src/stores/contract-store.ts
@@ -3,12 +3,12 @@ import { BigNumber } from 'ethers';
 import { DI, IContainer, ILogger, Registration } from 'aurelia';
 import { Erc20, TransferEvent } from 'models/generated/monetary/erc20/Erc20';
 import { Erc721 } from 'models/generated/monetary/erc721';
-import { INumberService, IServices, ITokenInfo, fromWei, toWei } from '../services';
+import { IContractService, tokenInfos } from 'services/contract';
+import { INumberService, ITokenInfo, fromWei, toWei } from '../services';
 import { Oracle } from 'models/generated/monetary/oracle';
 import { Reserve } from 'models/generated/monetary/reserve';
 import { Transaction } from 'models/transaction';
 import { Treasury } from 'models/generated/monetary/treasury';
-import { tokenInfos } from 'services/contract';
 export type IContractStore = ContractStore;
 export const IContractStore = DI.createInterface<IContractStore>('IContractStore');
 
@@ -17,7 +17,7 @@ export class ContractStore {
     container.register(Registration.singleton(IContractStore, ContractStore));
   }
   constructor(
-    @IServices private readonly services: IServices,
+    @IContractService private readonly contractService: IContractService,
     @ILogger private readonly logger: ILogger,
     @INumberService private readonly numberService: INumberService,
   ) {}
@@ -41,8 +41,8 @@ export class ContractStore {
     }
 
     const tokenContract = tokenInfo.id
-      ? this.services.contractService.getTokenContract(assetAddress, tokenInfo.id)
-      : this.services.contractService.getTokenContract(assetAddress); //get the ERC20 contract from the asset's address
+      ? this.contractService.getTokenContract(assetAddress, tokenInfo.id)
+      : this.contractService.getTokenContract(assetAddress); //get the ERC20 contract from the asset's address
 
     if (!oracleAddress) {
       if (tokenInfo.id) {
@@ -54,11 +54,11 @@ export class ContractStore {
     }
 
     if (!oracleAddress || BigNumber.from(oracleAddress).isZero()) return;
-    const oracleContract = this.services.contractService.getContract('Monetary', 'Oracle', oracleAddress) as Oracle; //get the oracle contract for the given oracle address
+    const oracleContract = this.contractService.getContract('Monetary', 'Oracle', oracleAddress) as Oracle; //get the oracle contract for the given oracle address
     const data = await oracleContract.getData(); // get the data from the oracle contract
 
     if (!data[1]) return; // if the oracleContract.getData() returns false don't use this token's data (according to Marvin G.)
-    tokenInfo.price = this.services.numberService.fromString(fromWei(data[0], 18)) ?? 0; //price comes back as undefined from getTokenInfoFromAddress so set it
+    tokenInfo.price = this.numberService.fromString(fromWei(data[0], 18)) ?? 0; //price comes back as undefined from getTokenInfoFromAddress so set it
 
     let tokenQuantity = toWei(1, 18); //all NFTs have a quantity of 1, so set the quantity to 1 initially
     let totalSupply: BigNumber | undefined;
@@ -85,7 +85,7 @@ export class ContractStore {
       type: assetType,
       totalSupply: totalSupply,
     };
-    asset.total = tokenInfo.id ? tokenInfo.price : (this.services.numberService.fromString(fromWei(asset.quantity, 18)) ?? 0) * tokenInfo.price;
+    asset.total = tokenInfo.id ? tokenInfo.price : (this.numberService.fromString(fromWei(asset.quantity, 18)) ?? 0) * tokenInfo.price;
     return asset;
   }
